refactor: moved createResource out of vault router
Moved createResource out of the vault router to the resource router as
it's own create route.

Also snuck in the userId field for the resource to allow querying
resources specifically from a user. Started implementing search in here
as well
